+++ modules/browser-node/src/services/store.ts
@@ -173,7 +173,7 @@ export class BrowserStore implements IEngineStore, IChainServiceStore {
     await this.db.transaction("rw", this.db.channels, this.db.transfers, async () => {
       // remove all "active" transfers
       const currActive = await this.getActiveTransfers(channelState.channelAddress);
-      // TODO: can we "unassociate" them without deleting them?
+      // TODO: can we "unassociate" them without deleting them? GH #431
       await this.db.transfers.bulkDelete(currActive.map((t) => t.transferId));
       // save channel
       await this.db.channels.put(channelState);