+++ modules/router/src/listener.ts
@@ -173,9 +173,6 @@ export async function setupListeners(
         (t) => t !== data.transfer.transferId,
       );
       if (res.isError) {
-        const { receiverAmount } = res.getError()?.context;
-        // TODO: if we did *not* cancel sender side transfer, should we
-        // increment?
         failedTransfer.inc({
           assetId,
           chainId,
@@ -446,7 +443,7 @@ export async function setupListeners(
   });
 
   nodeService.on(EngineEvents.DEPOSIT_RECONCILED, async (data) => {
-    // TODO: do we want this to be updated on withdrawal as well
+    // TODO: do we want this to be updated on withdrawal as well #442
     const channelRes = await nodeService.getStateChannel({ channelAddress: data.channelAddress });
     if (channelRes.isError) {
       logger.warn({ ...channelRes.getError()?.toJson() }, "Failed to get channel");
@@ -545,11 +542,13 @@ export async function setupListeners(
       recipient: _recipient,
       recipientChainId: _recipientChainId,
       recipientAssetId: _recipientAssetId,
+      receiveExactAmount: _receiveExactAmount,
     } = request.getValue();
 
     const recipient = _recipient ?? routerSigner.publicIdentifier;
     const recipientChainId = _recipientChainId ?? chainId;
     const recipientAssetId = _recipientAssetId ?? assetId;
+    const receiveExactAmount = _receiveExactAmount ?? false;
 
     const isSwap = recipientChainId !== chainId || recipientAssetId !== assetId;
     const supported = isSwap
@@ -691,8 +690,9 @@ export async function setupListeners(
       }
       recipientChannel = placeholder.getValue();
     }
-    const fee = await calculateFeeAmount(
+    const feeRes = await calculateFeeAmount(
       BigNumber.from(amount),
+      receiveExactAmount,
       assetId,
       senderChannel,
       recipientAssetId,
@@ -701,12 +701,12 @@ export async function setupListeners(
       routerSigner.publicIdentifier,
       logger,
     );
-    if (fee.isError) {
+    if (feeRes.isError) {
       await messagingService.respondToTransferQuoteMessage(
         inbox,
         Result.fail(
           new QuoteError(QuoteError.reasons.CouldNotGetFee, {
-            feeError: jsonifyError(fee.getError()!),
+            feeError: jsonifyError(feeRes.getError()!),
             recipient,
             recipientChainId,
             recipientAssetId,
@@ -718,15 +718,16 @@ export async function setupListeners(
       );
       return;
     }
+    const { fee, amount: quoteAmount } = feeRes.getValue();
     const quote = {
       assetId,
-      amount,
+      amount: quoteAmount.toString(),
       chainId,
       routerIdentifier: routerSigner.publicIdentifier,
       recipient,
       recipientChainId,
       recipientAssetId,
-      fee: fee.getValue().toString(),
+      fee: fee.toString(),
       expiry: (Date.now() + (getConfig().feeQuoteExpiry ?? DEFAULT_FEE_EXPIRY)).toString(), // valid for next 2 blocks
     };
     const toSign = hashTransferQuote(quote);