+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,20 @@
 
 ## Next Release
 
+## 0.2.0-beta.9
+
+- \[engine\] Add `transaction` to `WITHDRAWAL_RESOLVED` event
+- \[node\] Make `transaction` a non-optional return value
+
+## 0.2.0-beta.8
+
+- \[engine, router\] Allow users to specify if they want to receive an exact amount when requesting a quote
+
+## 0.2.0-beta.7
+
+- [engine] Fix transfer event balances
+- [engine] Allow Bob to specify submitting their own withdrawals
+
 ## 0.2.0-beta.6
 
 - \[engine\] Only error if fees > amount, not fees >= amount